--- src/custom/state/swapVCow/reducer.ts
@@ -1,16 +1,0 @@
-import { createReducer } from '@reduxjs/toolkit'
-import { SwapVCowStatus, setStatus } from './actions'
-
-export type SwapVCowState = {
-  status: SwapVCowStatus
-}
-
-export const initialState: SwapVCowState = {
-  status: SwapVCowStatus.INITIAL,
-}
-
-export default createReducer(initialState, (builder) =>
-  builder.addCase(setStatus, (state, { payload }) => {
-    state.status = payload
-  })
-)
